requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-multipart==0.0.6
jiwer==3.0.3
python-dotenv==1.0.0
numpy==2.2.6
pandas
joblib
scikit-learn